+++ engine/internal/httpapi/identity.go
@@ -1,13 +1,13 @@
 package httpapi
 
 import (
-	"net/http"
+	"context"
 
 	"k8s.io/client-go/rest"
 
 	"github.com/RajaSardar/kubebay/engine/internal/clusters"
 )
 
-func restConfigFor(r *http.Request, m *clusters.Manager, cluster string) (*rest.Config, error) {
-	return m.RestConfigWithIdentity(cluster, IdentityFromContext(r.Context()))
+func restConfigFor(ctx context.Context, m *clusters.Manager, cluster string) (*rest.Config, error) {
+	return m.RestConfigWithIdentity(cluster, clusters.IdentityFromContext(ctx))
 }